Added interface IApp, class ConsoleApp, class Container
