+++ TagsCloudContainer/Application/ConsoleApp.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Drawing.Imaging;
+using System.IO;
+using TagsCloudContainer.Application;
+
+namespace TagsCloudContainer
+{
+    public class ConsoleApp : IApp
+    {
+        private readonly IVisualisator _visualisator;
+        
+        public ConsoleApp(IVisualisator visualisator)
+        {
+            _visualisator = visualisator;
+        }
+
+        public void Run()
+        {
+            var projectDirectory = Directory.GetParent(Environment.CurrentDirectory).Parent.Parent.FullName;
+            _visualisator.Paint();
+            _visualisator.Save(projectDirectory + "\\Results", "Rectangles", ImageFormat.Png);
+        }
+    }
+}
