+++ TagsCloudContainer/Application/IApp.cs
@@ -1,0 +1,7 @@
+﻿namespace TagsCloudContainer.Application
+{
+    public interface IApp
+    {
+        public void Run();
+    }
+}
